temp: Use `NonNull::expose_provenance`
This commit is temporary. I'm not ready to try to make a PR with this
change, but I want to run CI with it on my fork (since my local machine
is still suffering from the issue I reported earlier). I'm also still
trying to make sure I understand the implications of the change.

I might be reading into it too much.

Signed-off-by: hashcatHitman <155700084+hashcatHitman@users.noreply.github.com>